+++ packages/next/src/index.client.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import {
   Var,
   Num,
@@ -6,33 +8,48 @@ import {
   T,
   Branch,
   Plural,
+  LocaleSelector,
+  useGT,
+  useTranslations,
+  useLocale,
+  useLocales,
+  useDefaultLocale,
+  useGTClass,
+  useLocaleProperties,
 } from 'gt-react/client';
+import {
+  gtProviderUseClientError,
+  txUseClientError,
+} from './errors/createErrors';
 
 // Mock <GTProvider> which throws an error
-export function GTProvider({}): React.JSX.Element {
-  throw new Error(
-    `You're attempting to import the Next.js <GTProvider> in a client component. ` +
-      `Are you sure you want to do this? It's better to import <GTProvider> in a file not marked 'use client' so that it can fetch translations on the server. ` +
-      `If you really need to put <GTProvider> on the client, import <GTClientProvider> from 'gt-next/client' instead (discouraged when using the Next.js App Router).`
-  );
+export function GTProvider() {
+  throw new Error(gtProviderUseClientError);
 }
 
-// Mock <TX> which throws an error
-export function Tx({}): React.JSX.Element {
-  throw new Error(
-    `You're attempting to use the <Tx> runtime translation component in a client component. ` +
-      `This is currently unsupported. Please use <T> with variables, ` +
-      `or make sure <Tx> rendered on the server only. `
-  );
+// Mock <Tx> which throws an error
+export function Tx() {
+  throw new Error(txUseClientError);
 }
 
 export {
-  // GTProvider
-  T, // Tx
+  T,
   Var,
   Num,
   Currency,
   DateTime,
   Branch,
   Plural,
+  LocaleSelector,
+  useGT,
+  useTranslations,
+  /**
+   * @deprecated Use useTranslations instead
+   */
+  useTranslations as useDict,
+  useLocale,
+  useLocales,
+  useDefaultLocale,
+  useGTClass,
+  useLocaleProperties,
 };