+++ packages/next/src/server-dir/buildtime/getTranslations.tsx
@@ -13,27 +13,28 @@ import {
   translationLoadingWarning,
 } from '../../errors/createErrors';
 import getI18NConfig from '../../config-dir/getI18NConfig';
-import getLocale from '../../request/getLocale';
+import { getLocale } from '../../request/getLocale';
 import {
   renderContentToString,
   splitStringToContent,
 } from 'generaltranslation';
 import { hashJsxChildren } from 'generaltranslation/id';
+import use from '../../utils/use';
 
 /**
- * Returns the dictionary access function `d()`, which is used to translate an item from the dictionary.
+ * Returns the dictionary access function t(), which is used to translate an item from the dictionary.
  *
  * @param {string} [id] - Optional prefix to prepend to the translation keys.
  * @returns {Function} A translation function that accepts a key string and returns the translated value.
  *
  * @example
- * const d = await getDict('user');
- * console.log(d('name')); // Translates item 'user.name'
+ * const t = await getTranslations('user');
+ * console.log(t('name')); // Translates item 'user.name'
  *
- * const d = await getDict();
- * console.log(d('hello')); // Translates item 'hello'
+ * const t = await getTranslations();
+ * console.log(t('hello')); // Translates item 'hello'
  */
-export default async function getDict(
+export async function getTranslations(
   id?: string
 ): Promise<(id: string, options?: DictionaryTranslationOptions) => string> {
   // ---------- SET UP ---------- //
@@ -58,16 +59,16 @@ export default async function getDict(
 
   const renderSettings = I18NConfig.getRenderSettings();
 
-  // ---------- THE d() METHOD ---------- //
+  // ---------- THE t() METHOD ---------- //
 
   /**
-   * @description A function that translates a dictionary entry based on its `id` and options.
+   * @description A function that translates a dictionary entry based on its id and options.
    * @param {string} id The identifier of the dictionary entry to translate.
    * @param {DictionaryTranslationOptions} options
    * @returns The translated version of the dictionary entry.
    *
    * @example
-   * d('greetings.greeting1'); // Translates item in dictionary under greetings.greeting1
+   * t('greetings.greeting1'); // Translates item in dictionary under greetings.greeting1
    *
    * @example
    * // dictionary entry
@@ -78,9 +79,9 @@ export default async function getDict(
    * }
    *
    * // Translates item in dictionary under greetings.greeting2 and replaces {name} with 'John'
-   * d('greetings.greeting2', { variables: { name: 'John' } });
+   * t('greetings.greeting2', { variables: { name: 'John' } });
    */
-  const d = (
+  const t = (
     id: string,
     options: DictionaryTranslationOptions = {}
   ): string => {
@@ -190,5 +191,22 @@ export default async function getDict(
     return renderContent(source, [defaultLocale]);
   };
 
-  return d;
+  return t;
+}
+
+/**
+ * Returns the dictionary access function t(), which is used to translate an item from the dictionary.
+ *
+ * @param {string} [id] - Optional prefix to prepend to the translation keys.
+ * @returns {Function} A translation function that accepts a key string and returns the translated value.
+ *
+ * @example
+ * const t = useTranslations('user');
+ * console.log(t('name')); // Translates item 'user.name'
+ *
+ * const t = useTranslations();
+ * console.log(t('hello')); // Translates item 'hello'
+ */
+export function useTranslations(id?: string) {
+  return use(getTranslations(id));
 }