+++ packages/next/src/variables/DateTime.tsx
@@ -1,35 +1,19 @@
 import { formatDateTime } from 'generaltranslation';
 import getI18NConfig from '../config-dir/getI18NConfig';
-import getLocale from '../request/getLocale';
+import { getLocale, useLocale } from '../request/getLocale';
 
-/**
- * The `<DateTime>` component renders a formatted date or time string, allowing customization of the name, default value, and formatting options.
- * It utilizes the current locale and optional format settings to display the date.
- *
- * @example
- * ```jsx
- * <DateTime>
- *    {new Date()}
- * </DateTime>
- * ```
- *
- * @param {any} [children] - Optional content (typically a date) to render inside the component.
- * @param {Intl.DateTimeFormatOptions} [options={}] - Optional formatting options for the date, following `Intl.DateTimeFormatOptions` specifications.
- * @returns {Promise<React.JSX.Element>} The formatted date or time component.
- */
-async function DateTime({
+function DateTime({
   children,
-  name,
   locales,
   options = {},
 }: {
-  children?: any;
+  children?: React.ReactNode;
   name?: string;
   options?: Intl.DateTimeFormatOptions; // Optional formatting options for the date
   locales?: string[];
-}): Promise<React.JSX.Element> {
+}): React.JSX.Element {
   if (!locales) {
-    locales = [await getLocale(), getI18NConfig().getDefaultLocale()];
+    locales = [useLocale(), getI18NConfig().getDefaultLocale()];
   }
 
   let final;